refactor: move nodeOpsUtils to utils
